BUGFIX: Multi-letter-options prevent stacking
With option 'x' and 'abc' you can stack them as '-xabc'
which is the same as '-x -abc'.

However they stacked as well as '-abcx' which was confusing.
This now gives 'unknown option -abcx' which is much better.

Also following case with LOPTs with arguments was fixed:
'-loptARG' was '-lopt' with arg 'RG' (with 'A' silently swallowed)
which was very confusing, too.  Also '-lopt.ARG' was possible.
Both now do no more work.  Both now are 'unknown option'.