tino_getopt(): new meaning of "opt=VAL helptext"
Previously this would have been option '-opt=VAL'.
Now this is parsed as option '-opt', so parsing of option
name stops at the first '='.  If no _DIRECT is present,
this means you can give it as '-opt=ARG' or '-opt ARG'.

Note that '-opt= helptext' means, that the option is '-opt='.
Hence you can give '-opt=ARG' or '-opt= ARG' but not '-opt ARG'.

Best is to use _DIRECT in this case, such that '-opt= ARG' is
interpreted as '-opt=' without arg, (mostly) followed by '-- ARG'.