Actually the last commit was a BUGFIX
On 64 bit platforms an internal buffer could overflow.
This could be exploited if backup names were carefully crafted.

However only numbers can be created this way, so it is unlikely
that something worse happens than a probably program crash.

But who knows?  So extended the buffer even more to avoid this BUG
until 512 bit "long"s, which hopefully never happens.