sock.h: IPv6 address resolution works now
Today I learned that gethostbyname is IPv4 only.  Sort of.
It can only do IPv6 if the whole system is switched to IPv6.
Which is nuts.  The app must be able to use all available protocols.

This now is supported with following types:

host:port	defaults to IPv4, fallback to IPv6
[host]:port	IPv6 only.  IPv4 hosts use IPv4 within IPv6
ip:port		IPv4 for IPv4-IPs, IPv6 for IPv6-IPs
[ip]:port	ditto

As before we can have Unix Domain (without : or starting with . or /)
and Abstract Linux sockets (starting with @).