Apparently #pragma is completely broken
tino/getopt.h:1337:9: error: expected expression before 'case'
 1337 |         case 'w':       f *= 7ull;                              /* Week */
      |         ^~~~

I suspect this is only because before it there is a #pragma!

So let's see if this change fixes this compiler bug.

With all this hassle the only thing I can say is, that the
pragma and warning concept of the compilers are now is at a stage
where it can be considered completely broken and harmful.